vendor/rsc.io/pdf: import rsc.io/pdf library
This will be needed by some of the arch subdirs.
It obviously shouldn't be exposed for import,
so it must be vendor or internal.

Using vendor here is a bit of an experiment in how we might
use vendor in our own subrepos. Since the subrepo is new it
seems a reasonable test case.

For golang/go#12179.

Change-Id: Ibe7596903c54e120dc585f146c221d996dde5808
Reviewed-on: https://go-review.googlesource.com/13968
Reviewed-by: Ian Lance Taylor <iant@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>