vendor: delete
The vendor directory was added in CL 13968, as part of the Go 1.5
vendor experiment (golang.org/s/go15vendor), before the module mode
was in existence to achieve these properties.

The version of the rsc.io/pdf dependency can be determined from the
go.mod file, resulting in reproducible builds.

The content of the rsc.io/pdf dependency can be fetched from origin,
proxy.golang.org, or another proxy, resulting in available builds.

The checksum of the rsc.io/pdf dependency is written down in the
go.sum file, as well as in sum.golang.org, resulting in secure builds.

I am not aware of a reason for the vendor directory to exist in this
repository anymore, so delete it in order to have fewer files.

Change-Id: Ib97b8ff164f9f5e6565cf5b49fb794f3e4aab43c
Reviewed-on: https://go-review.googlesource.com/c/arch/+/222539
Reviewed-by: Bryan C. Mills <bcmills@google.com>