x86/xeddata: rewrite Reader to use new readLines iterator
The parsing in readLines is simpler and more robust. It also includes
line information (which would have been pretty annoying to add to the
old parser), so this is a step toward adding position information to
Inst.

One downside of this is that Reader.Read has to use a pull iterator.
However, as far as I can tell, the only callers of this were
Reader.ReadAll and WalkInsts, both of which we rewrite to use
readObjects directly.

Change-Id: I4ca58c877fbfd5295209aea31999c8abd6876f17
Reviewed-on: https://go-review.googlesource.com/c/arch/+/656238
Auto-Submit: Austin Clements <austin@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>