riscv64: add support for RVV 1.0 instructions
Add support for the dissassembly of RVV 1.0 instructions and a lot of
test cases.  The test cases are based on the test cases found in
go/src/cmd/asm/internal/asm/testdata/riscv64.s in the main go repo.

We rename arg_rs1_amo, an argument that represents a pointer without
an offset, to arg_rs1_ptr so it can be reused by the vector
instructions.

Change-Id: Icd7ec2e853fe03f94aac35698f44e1aeaac2edae
Reviewed-on: https://go-review.googlesource.com/c/arch/+/670876
Reviewed-by: Meng Zhuo <mengzhuo1203@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Joel Sing <joel@sing.id.au>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Mark Freeman <markfreeman@google.com>