internal/unify: fix parsing of empty string nodes
Currently, if a node is just "", we incorrectly parse that as a regexp
instead of an exact string. This happens because we build it into the
regexp "\A(?:)\z" and then test if it's "exact" by asking if it has a
"literal prefix". But for the empty string and exactly the empty
string, it's literal but has no prefix, so this check fails and we
treat this as a regexp instead of an exact string.

Fix this by special-casing "".

I believe this bug is harmless for unification, since either way it
describes the same set of strings, but it affects any "exactness" test
and affects YAML round-tripping.

Change-Id: I3223ef9a27c3cb6bfd2f5a0be9a0b7b71059840f
Reviewed-on: https://go-review.googlesource.com/c/arch/+/689479
Reviewed-by: David Chase <drchase@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Austin Clements <austin@google.com>