internal/simdgen: make sure that output is based on sorted data
there was still some variation, this may not be "the best"
order in all cases, but it is definitely better than no order,
and we can tweak individual files as we decide it is suitable.

this does not change the current generated files, but that
turns out to be just luck.

Change-Id: I38c6ac72f69b9d29c71de3250985cff8b7fcd677
Reviewed-on: https://go-review.googlesource.com/c/arch/+/695335
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Junyang Shao <shaojunyang@google.com>