riscv64: fix and update decoding of MOVD and MOVF
Change https://go-review.googlesource.com/c/go/+/703216 enhances
MOVD and MOVF so that they can be used to initialise a floating
point register to 0.0 and also to copy the binary representations
of floating point values between integer and floating point registers.

We update the riscv64 plan9 disassembler to recognise these cases
and output the appropriate MOVD/MOVF instructions.

We also fix a bug in the plan9 decoder which results in the
arguments for the MOVD and MOVF instructions that represent loads
and stores being output in the wrong order.

Change-Id: I5c3561437f5a0c936767497de6e22c976ce17cff
Reviewed-on: https://go-review.googlesource.com/c/arch/+/704195
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Joel Sing <joel@sing.id.au>
Reviewed-by: Junyang Shao <shaojunyang@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Meng Zhuo <mengzhuo1203@gmail.com>